# * existing of ENV key should be sufficient to enable