--- a/lib/utils/transpile-es6.js
+++ b/lib/utils/transpile-es6.js
@@ -6,11 +6,16 @@ var babel = require('babel-core');
 var Cache = require('sync-disk-cache');
 var crypto = require('crypto');
 
-var SHOULD_CACHE = process.env.DISK_CACHE;
+var SHOULD_CACHE = 'DISK_CACHE' in process.env;
 var cache = new Cache('emberjs-build-es6');
 
-if (process.env.PURGE_DISK_CACHE) {
+if ('PURGE_DISK_CACHE' in process.env) {
+  console.log('[purging] persistent es6 transpile cache');
   cache.clear();
+}
+
+if (SHOULD_CACHE) {
+  console.log('[enabled] persistent es6 transpile cache');
 }
 
 module.exports = function(tree, description, options) {