--- a/lib/utils/transpile-es6.js
+++ b/lib/utils/transpile-es6.js
@@ -15,7 +15,7 @@ if ('PURGE_DISK_CACHE' in process.env) {
 }
 
 if (SHOULD_CACHE) {
-  console.log('[enabled] persistent es6 transpile cache');
+  console.log('[enabled] persistent es6 transpile cache path:', cache.root);
 }
 
 module.exports = function(tree, description, options) {