* log if the cache is enabled