@stefanpenner 2015-06-04
before caching:

```
± % DISABLE_ES3='true' DISABLE_JSHINT=true  DISABLE_DEREQUIRE='true' ember s --port 1338                                                      !15361
Future versions of Ember CLI will not support v0.10.35. Please update to Node 0.12 or io.js.
version: 0.2.5

A new version of ember-cli is available (0.2.7). To install it, type ember update.
Livereload server on port 32867
Serving on http://localhost:1338/
[BABEL] Note: The code generator has deoptimised the styling of "ember/tests/routing/basic_test.js" as it exceeds the max of "100KB".

Build successful - 49471ms.

Slowest Trees                                 | Total
----------------------------------------------+---------------------
ES6 Modules                                   | 21973ms
ES6 Modules                                   | 9519ms
ES6 Modules                                   | 3932ms

Slowest Trees (cumulative)                    | Total (avg)
----------------------------------------------+---------------------
ES6 Modules (15)                              | 41958ms (2797 ms)
 (30)                                         | 4417ms (147 ms)
```

after:
```
± % DISABLE_ES3='true' DISABLE_JSHINT=true DISK_CACHE='true' DISABLE_DEREQUIRE='true' ember s --port 1338                                     !15361
Future versions of Ember CLI will not support v0.10.35. Please update to Node 0.12 or io.js.
version: 0.2.5

A new version of ember-cli is available (0.2.7). To install it, type ember update.
[enabled] persistent es6 transpile cache path: /var/folders/tc/24p44f8x48jc0jft9nqcdqm4002qv6/T/emberjs-build-es6
Livereload server on port 32867
Serving on http://localhost:1338/

Build successful - 8913ms.

Slowest Trees                                 | Total
----------------------------------------------+---------------------
ES6 Modules                                   | 969ms
                                              | 728ms
                                              | 584ms
SourcemapConcat                               | 531ms
                                              | 496ms
                                              | 492ms
UseStrictRemover                              | 472ms

Slowest Trees (cumulative)                    | Total (avg)
----------------------------------------------+---------------------
 (30)                                         | 4416ms (147 ms)
ES6 Modules (15)                              | 1603ms (106 ms)
UseStrictRemover (9)                          | 738ms (82 ms)
SourcemapConcat (3)                           | 590ms (196 ms)



```