@rwjblue 2015-06-04
Released as 0.1.4.